support hiding plugins (depends on settings) (#21)
* s(wip): this would work if RGL was better

* hiding works (driven by hidden: true}
