+++ src/app/helpers/constants.ts
@@ -1,0 +1,3 @@
+
+// the constant for the name of the install lockfile
+export const pluginInstalledLockFileName = 'overlayed-install.lock'
